[CMake] Split out WebPageProxy+Swift.cpp https://bugs.webkit.org/show_bug.cgi?id=316471 rdar://178872817
Reviewed by NOBODY (OOPS!).

This shrinks the single-core long tail at the end the build.
WebPageProxy+Swift.cpp needs to wait for Swift to build a
bunch of modules, but WebPageProxy.cpp can compile earlier
without waiting.

* Source/WebKit/CMakeLists.txt: Add WebPageProxy+Swift.cpp

* Source/WebKit/Sources.txt:
* Source/WebKit/UIProcess/WebPageProxy+Swift.cpp: Split out
from WebPageProxy.cpp, cutting out the parts that use Swift.
* Source/WebKit/UIProcess/WebPageProxy.cpp:

* Source/WebKit/UIProcess/WebPageProxyInlines.h: New file
to hold shared send<> function template instantiations.

* Source/WebKit/WebKit.xcodeproj/project.pbxproj: Update the
Xcode build to build both files.
